+++ client/pom.xml
@@ -3,7 +3,7 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>java-basic-http</artifactId>
+        <artifactId>http</artifactId>
         <groupId>ru.fisunov.http</groupId>
         <version>1.0</version>
     </parent>
