+++ client/src/main/java/ru/otus/http/client/MainApplication.java
