+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>ru.fisunov.http</groupId>
-    <artifactId>java-basic-http</artifactId>
+    <artifactId>http</artifactId>
     <packaging>pom</packaging>
     <version>1.0</version>
