+++ server/src/main/java/ru/otus/http/server/Request.java
@@ -1,4 +1,7 @@
-package ru.fisunov.http.server;
+package ru.otus.http.server;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -8,6 +11,7 @@ public class Request {
     private String uri;
     private String raw;
     private Map<String, String> params;
+    private static final Logger logger = LogManager.getLogger(Request.class);
 
     public String getUri() {
         return uri;
@@ -47,9 +51,10 @@ public class Request {
     }
 
     public void show() {
-        System.out.println("Запрос:");
-        System.out.println("uri: " + uri);
-        System.out.println("params: " + params);
+        logger.info("Получен запрос:");
+        logger.info("Запрос:");
+        logger.info("uri: " + uri);
+        logger.info("params: " + params);
     }
 
     public String getParam(String key) {
